Move record events to analytics file
