+++ src/applications/facility-locator/components/search-results-items/VaFacilityResult.jsx
@@ -2,7 +2,8 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import LocationPhoneLink from './common/LocationPhoneLink';
 import LocationDirectionsLink from './common/LocationDirectionsLink';
-import { isVADomain, recordResultClickEvents } from '../../utils/helpers';
+import { isVADomain } from '../../utils/helpers';
+import { recordResultClickEvents } from '../../utils/analytics';
 import { Link } from 'react-router';
 import { OperatingStatus } from '../../constants';
 import LocationAddress from './common/LocationAddress';
